smart_config

Module _docs

Available on crate feature _docs only.

Additional library documentation

This module and submodules contains additional docs on how to use the smart-config library.

Derive macros reference

Derive macro examples

Combining configuration sources

Links to “main” documentation

For completeness, here are some advanced topics documented in the corresponding modules of the library:

Deserialization overview

Config validation

Parameter fallbacks

String pattern matching

Testing configurations

Modules

derive_examples

Derive macro examples

derive_ref

Derive macros reference

sources

Parsing configurations from multiple sources