smart_config::de

Trait DeserializeParam

pub trait DeserializeParam<T>:
    Debug
    + Send
    + Sync
    + 'static {
    const EXPECTING: BasicTypes;

    // Required methods
    fn deserialize_param(
        &self,
        ctx: DeserializeContext<'_>,
        param: &'static ParamMetadata,
    ) -> Result<T, ErrorWithOrigin>;
    fn serialize_param(&self, param: &T) -> Value;

    // Provided method
    fn describe(&self, description: &mut TypeDescription) { ... }
}

Deserializes a parameter of the specified type.

Implementations

Basic implementations

Serde allows deserializing any type implementing serde::Deserialize.

TimeUnit deserializes Duration from a numeric value that has the specified unit of measurement

SizeUnit similarly deserializes ByteSize

WithUnit deserializes Durations / ByteSizes as an integer + unit of measurement (either in a string or object form).

Decorators

Optional decorates a deserializer for T turning it into a deserializer for Option<T>

WithDefault adds a default value used if the input is missing

Delimited allows deserializing arrays from a delimited string (e.g., comma-delimited)

OrString allows to switch between structured and string deserialization

Required Associated Constants

const EXPECTING: BasicTypes

Describes which parameter this deserializer is expecting.

Required Methods

fn deserialize_param( &self, ctx: DeserializeContext<'_>, param: &'static ParamMetadata, ) -> Result<T, ErrorWithOrigin>

Performs deserialization given the context and param metadata.

Errors

Returns an error if a param cannot be deserialized, e.g. if it has an incorrect type.

fn serialize_param(&self, param: &T) -> Value

Serializes the provided parameter to the JSON model.

Serialization is considered infallible (serde_json serialization may fail on recursive or very deeply nested data types; please don’t use such data types for config params).

Provided Methods

fn describe(&self, description: &mut TypeDescription)

Additional info about the deserialized type, e.g., extended description.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.