Trait smart_config::de::WellKnown

pub trait WellKnown: 'static {
    const DE: &'static dyn DeserializeParam<Self>;
}

Parameter type with well-known deserializer.

Conceptually, this means that the type is known to behave well when deserializing data from a Value (ordinarily, using serde::Deserialize).

Required Associated Constants

const DE: &'static dyn DeserializeParam<Self>

Standard deserializer for the param.

Object Safety

This trait is not object safe.