Trait smart_config::de::WellKnownOption

pub trait WellKnownOption: WellKnown { }

Marker trait for types that use a conventional Optional deserializer for Option<Self>.

It’s usually sound to implement this trait for custom types together with WellKnown, unless:

The type needs custom null coercion logic (e.g., coercing some structured values to null). In this case, implement CustomKnownOption instead. Note that WellKnownOption is tied to it via a blanket implementation.

It doesn’t make sense to have optional type params.

Object Safety

This trait is not object safe.