Macro smart_config::Custom

macro_rules! Custom {
    ($typ:ty; *) => { ... };
    ($typ:ty; $($expecting:tt),+ $(,)?) => { ... };
}

Constructor of Custom types / instances.

The macro accepts a deserialized type followed by a comma-separated list of expected basic types from the following set: bool, int, float, str, array, object. As a shortcut, Custom![_; *] signals to accept any input.

Examples

use serde::Deserialize;
use smart_config::de::Custom;

/// Deserializer that accepts a string and maps it to its length.
const LEN_DESERIALIZER: Custom![usize; str] = Custom![_; str](|ctx, param| {
    let de = ctx.current_value_deserializer(param.name)?;
    Ok(String::deserialize(de)?.len())
});