smart_config::metadata

Struct ParamMetadata

pub struct ParamMetadata {
    pub name: &'static str,
    pub aliases: &'static [(&'static str, AliasOptions)],
    pub help: &'static str,
    pub rust_field_name: &'static str,
    pub rust_type: RustType,
    pub expecting: BasicTypes,
    pub tag_variant: Option<&'static ConfigVariant>,
    /* private fields */
}

Metadata for a specific configuration parameter.

Fields
name: &'static str

Canonical param name in the config sources. Not necessarily the Rust field name!

aliases: &'static [(&'static str, AliasOptions)]

Param aliases.

help: &'static str

Human-readable param help parsed from the doc comment.

rust_field_name: &'static str

Name of the param field in Rust code.

rust_type: RustType

Rust type of the parameter.

expecting: BasicTypes

Basic type(s) expected by the param deserializer.

tag_variant: Option<&'static ConfigVariant>

Tag variant in the enclosing ConfigMetadata that enables this parameter. None means that the parameter is unconditionally enabled.

Implementations

impl ParamMetadata

pub fn default_value(&self) -> Option<Box<dyn Any>>

Returns the default value for the param.

pub fn default_value_json(&self) -> Option<Value>

Returns the default value for the param serialized into JSON.

pub fn example_value_json(&self) -> Option<Value>

Returns the example value for the param serialized into JSON.

pub fn type_description(&self) -> TypeDescription

Returns the type description for this param as provided by its deserializer.