Struct smart_config::ParseError

pub struct ParseError { /* private fields */ }

Config parameter deserialization errors.

Implementations

impl ParseError

pub fn inner(&self) -> &Error

Returns the wrapped error.

pub fn path(&self) -> &str

Returns an absolute path on which this error has occurred.

pub fn origin(&self) -> &ValueOrigin

Returns an origin of the value deserialization of which failed.

pub fn validation(&self) -> Option<&str>

Returns human-readable description of the failed validation, if the error is caused by one.

pub fn config(&self) -> &'static ConfigMetadata

Returns metadata for the failing config.

pub fn param(&self) -> Option<&'static ParamMetadata>

Returns metadata for the failing parameter if this error concerns a parameter. The parameter is guaranteed to be contained in Self::config().