Trait smart_config::de::WellKnown

pub trait WellKnown: 'static + Sized {
    type Deserializer: DeserializeParam<Self>;

    const DE: Self::Deserializer;
}

Parameter type with well-known deserializer.

Conceptually, this means that the type is known to behave well when deserializing data from a Value (ordinarily, using serde::Deserialize).

Implementations

Basic well-known types include:

bool

String

PathBuf

Signed and unsigned integers, including non-zero variants

f32, f64

These types use Serde deserializer.

WellKnown is also implemented for more complex types:

Rust type	Deserializer	Expected JSON

Duration	WithUnit	string or object

ByteSize	WithUnit	string or object

Option	Optional†	value, or null, or nothing

Vec, [_; N], HashSet, BTreeSet	Repeated	array

HashMap, BTreeMap	RepeatedEntries	object

† Options handling can be customized via WellKnownOption or CustomKnownOption traits.

Required Associated Types

type Deserializer: DeserializeParam<Self>

Type of the deserializer used for this type.

Required Associated Constants

const DE: Self::Deserializer

Deserializer instance.

Object Safety

This trait is not object safe.