smart_config::de

Trait WellKnown

pub trait WellKnown: 'static + Sized {
    type Deserializer: DeserializeParam<Self>;

    const DE: Self::Deserializer;
}

Parameter type with well-known deserializer.

Conceptually, this means that the type is known to behave well when deserializing data from a Value (ordinarily, using serde::Deserialize).

Implementations

Basic well-known types include:

bool

String

PathBuf

Signed and unsigned integers, including non-zero variants

f32, f64

These types use Serde deserializer. Integer types up to and including 64-bit ones only permit integer input. In contrast, u128, i128 and their non-zero counterparts permit either integer or string input because the serde_json object model cannot represent large integers (>u64::MAX or <i64::MIN). Importantly, string representation must be used for such large integers because otherwise they will be converted (with precision loss!) to f64 by internal serde_json logic before they ever enter the smart-config library.

WellKnown is also implemented for more complex types:

Rust type	Deserializer	Expected JSON

Duration	WithUnit	string or object

ByteSize	WithUnit	string or object

Option	Optional†	value, or null, or nothing

Vec, [_; N], HashSet, BTreeSet	Repeated	array

HashMap, BTreeMap	RepeatedEntries	object

† Options handling can be customized via WellKnownOption or CustomKnownOption traits.

Required Associated Constants

const DE: Self::Deserializer

Deserializer instance.

Required Associated Types

type Deserializer: DeserializeParam<Self>

Type of the deserializer used for this type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.