smart_config

Module metadata

Configuration metadata.

Structs

AliasOptions

Options for a param or config alias.

BasicTypes

Set of one or more basic types in the JSON object model.

ConfigMetadata

Metadata for a configuration (i.e., a group of related parameters).

ConfigTag

Information about a config tag.

ConfigVariant

Variant of a ConfigTag.

NestedConfigMetadata

Mention of a nested configuration within a configuration.

ParamMetadata

Metadata for a specific configuration parameter.

RustType

Representation of a Rust type.

TypeDescription

Human-readable description for a Rust type used in configuration parameter (Boolean value, integer, string etc.).

Enums

EtherUnit

Unit of ether amount measurement.

SizeUnit

Unit of byte size measurement.

TimeUnit

Unit of time measurement.

UnitOfMeasurement

General unit of measurement.