smart_config

Module testing

Testing tools for configurations.

Structs

Tester

Test case builder that allows configuring deserialization options etc.

Functions

test

Tests config deserialization from the provided sample. Takes into account param aliases, performs sample preprocessing etc.

test_complete

Tests config deserialization ensuring that all declared config params are covered.

test_minimal

Tests config deserialization ensuring that only required config params are covered. This is useful to detect new required params (i.e., backward-incompatible changes).