Trait smart_config::ExampleConfig

pub trait ExampleConfig {
    // Required method
    fn example_config() -> Self;
}

Provides an example for this configuration. The produced config can be used in tests etc.

For struct configs, this can be derived via the corresponding proc macro.

Required Methods

fn example_config() -> Self

Constructs an example configuration.

Object Safety

This trait is not object safe.