smart_config

Trait ExampleConfig

pub trait ExampleConfig {
    // Required method
    fn example_config() -> Self;
}

Provides an example for this configuration. The produced config can be used in tests etc.

For struct configs, this can be derived via the corresponding proc macro.

Required Methods

fn example_config() -> Self

Constructs an example configuration.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.