smart_config::validation

Trait Validate

pub trait Validate<T: ?Sized>:
    'static
    + Send
    + Sync {
    // Required methods
    fn describe(&self, formatter: &mut Formatter<'_>) -> Result;
    fn validate(&self, target: &T) -> Result<(), ErrorWithOrigin>;
}

Generic post-validation for a configuration parameter or a config.

Implementations

Validations are implemented for the following types:

NotEmpty. Validates that a string or a collection, such as Vec, is not empty.

Range, RangeInclusive etc. Validates whether the type is within the provided bounds.

Required Methods

fn describe(&self, formatter: &mut Formatter<'_>) -> Result

Describes this validation.

Errors

Should propagate formatting errors.

fn validate(&self, target: &T) -> Result<(), ErrorWithOrigin>

Validates a parameter / config.

Errors

Should return an error if validation fails.